requirements.txt
requests
locustio
Faker
BeautifulSoup
urllib3
bokeh
numpy
matplotlib